frontend performance improved. And admin tools implemented.

+++ application/api_v1/users.py
@@ -7,12 +7,12 @@
     users API module
 
 """
-from flask import request, Blueprint
-from flask_security.utils import encrypt_password, verify_password, logout_user
+from flask import request, Blueprint, after_this_request, current_app
+from flask_security.utils import encrypt_password, logout_user
 from application.api_v1 import route, login_required
 from application.services import users
 from application.core import gravatar
-from application.form.users import UpdateForm
+from application.form.users import UpdateForm, UpdatePasswordForm
 from werkzeug.datastructures import MultiDict
 
 bp = Blueprint('users', __name__, url_prefix='/users')
@@ -25,6 +25,18 @@ def find_one(user_id):
     :param user_id: user's id(INT)
     :return: user, is_current, is_following, followers_cnt, following_cnt
     """
+
+    @after_this_request
+    def empty_temp(response):
+        users.reset_json_temp()
+        users.reset_json_hidden()
+        return response
+
+    if request.values.get('extra', None) is not None:
+        users.set_json_temp(str(request.values.get('extra')).split(','))
+        if not users.is_admin():
+            users.set_json_hidden(['password', 'last_login_ip', 'current_login_ip', 'roles'])
+
     user = users.get_or_404(user_id)
     return {'user': user, "is_current": users.is_current(user), "is_following": users.is_following(user),
             "followers_cnt": users.followers_count(user), "following_cnt": users.following_count(user),
@@ -38,9 +50,21 @@ def find_all():
 
     :return: users, has_next, has_prev, current_page
     """
-    current_page = int(request.values.get('current_page', 1))
-    user_items, has_next, has_prev = users.paginate(current_page)
-    return {'users': user_items, 'has_next': has_next, 'has_prev': has_prev, 'current_page': current_page}
+    page = int(request.values.get('page', 1))
+    user_items, has_prev, has_next = users.paginate(page)
+
+    @after_this_request
+    def empty_temp(response):
+        users.reset_json_temp()
+        users.reset_json_hidden()
+        return response
+
+    if request.values.get('extra', None) is not None:
+        users.set_json_temp(str(request.values.get('extra')).split(','))
+        if not users.is_admin():
+            users.set_json_hidden(['password', 'last_login_ip', 'current_login_ip', 'roles'])
+
+    return {'users': user_items, 'has_prev': has_prev, 'has_next': has_next, 'current_page': page}
 
 
 @route(bp, '/<user_id>', methods=['PUT'])
@@ -48,7 +72,19 @@ def find_all():
 def update(user_id):
     """
     Update user information.
+    def test_get_current_user(self):
+        r = self.jget('/users')
+        self.assertOkJson(r)
 
+    def test_get_user(self):
+        r = self.jget('/users/%s' % self.user.id)
+        self.assertOkJson(r)
+
+    def test_delete_user(self):
+
+        print("auth : %s")
+        r = self.xdelete('/users/%s' % self.user.id)
+        self.assertOkJson(r)
     @login_required
 
     :param user_id:
@@ -56,18 +92,25 @@ def update(user_id):
     """
 
     if request.form:
-        form = UpdateForm(MultiDict(request.form))
+        if request.form.get('newPassword') is not None:
+            form = UpdatePasswordForm(MultiDict(request.form))
+        else:
+            form = UpdateForm(MultiDict(request.form))
     else:
         form = UpdateForm()
-
     user = users.get_or_404(user_id)
-    password_incorrect = False
-    if form.validate_on_submit():
-        user.password = encrypt_password(request.form.get('newPassword'))
-        users.save(user)
+    if users.is_admin() or users.user_equals_me(user):
+        password_incorrect = False
+        if form.validate_on_submit():
+            if isinstance(form, UpdatePasswordForm):
+                user.password = encrypt_password(request.form.get('newPassword'))
+            # user.active = request.form.get('active')
+            users.update(user, **request.form)
+        else:
+            password_incorrect = True
+        return {'user': user, 'password_incorrect': password_incorrect}
     else:
-        password_incorrect = True
-    return {'user': user, 'password_incorrect': password_incorrect}
+        return {}, 401
 
 
 @route(bp, '/<user_id>', methods=['DELETE'])
@@ -82,8 +125,14 @@ def destroy(user_id):
     :return:
     """
     user = users.get_or_404(user_id)
-    users.delete_user(user)
-    logout_user()
-    return {}
-
+    if users.user_equals_me(user):
+        users.delete_user(user)
+        if not current_app.testing:
+            logout_user()
+        return {}
+    if users.is_admin():
+        users.delete_user(user)
+        return {}
+    else:
+        return {}, 401
